--- app/src/code-block/icons.tsx
@@ -1,86 +1,0 @@
-import React from 'react';
-
-export interface IconProps extends React.HTMLAttributes<SVGElement> {
-  size?: number;
-  color?: string;
-}
-
-export function Copy({ size = 14, color = 'currentColor', className = 'inline', ...props }: IconProps) {
-  return (
-    <svg
-      {...props}
-      width={size}
-      height={size}
-      className={className}
-      viewBox='0 0 15 15'
-      xmlns='http://www.w3.org/2000/svg'
-    >
-      <path
-        d='M1 9.50006C1 10.3285 1.67157 11.0001 2.5 11.0001H4L4 10.0001H2.5C2.22386 10.0001 2 9.7762 2 9.50006L2 2.50006C2 2.22392 2.22386 2.00006 2.5 2.00006L9.5 2.00006C9.77614 2.00006 10 2.22392 10 2.50006V4.00002L11 4.00002V2.50006C11 1.67163 10.3284 1.00006 9.5 1.00006H2.5C1.67157 1.00006 1 1.67163 1 2.50006V9.50006ZM5.5 4.00006C4.67157 4.00006 4 4.67163 4 5.50006V12.5001C4 13.3285 4.67157 14.0001 5.5 14.0001H12.5C13.3284 14.0001 14 13.3285 14 12.5001V5.50006C14 4.67163 13.3284 4.00006 12.5 4.00006H5.5ZM5 5.50006C5 5.22392 5.22386 5.00006 5.5 5.00006H12.5C12.7761 5.00006 13 5.22392 13 5.50006V12.5001C13 12.7762 12.7761 13.0001 12.5 13.0001H5.5C5.22386 13.0001 5 12.7762 5 12.5001V5.50006Z'
-        fill={color}
-        fillRule='evenodd'
-        clipRule='evenodd'
-      />
-    </svg>
-  );
-}
-
-export function Download({ size = 14, color = 'currentColor', className = 'inline', ...props }: IconProps) {
-  return (
-    <svg
-      {...props}
-      width={size}
-      height={size}
-      className={className}
-      viewBox='0 0 15 15'
-      xmlns='http://www.w3.org/2000/svg'
-    >
-      <path
-        d='M7.50005 1.04999C7.74858 1.04999 7.95005 1.25146 7.95005 1.49999V8.41359L10.1819 6.18179C10.3576 6.00605 10.6425 6.00605 10.8182 6.18179C10.994 6.35753 10.994 6.64245 10.8182 6.81819L7.81825 9.81819C7.64251 9.99392 7.35759 9.99392 7.18185 9.81819L4.18185 6.81819C4.00611 6.64245 4.00611 6.35753 4.18185 6.18179C4.35759 6.00605 4.64251 6.00605 4.81825 6.18179L7.05005 8.41359V1.49999C7.05005 1.25146 7.25152 1.04999 7.50005 1.04999ZM2.5 10C2.77614 10 3 10.2239 3 10.5V12C3 12.5539 3.44565 13 3.99635 13H11.0012C11.5529 13 12 12.5528 12 12V10.5C12 10.2239 12.2239 10 12.5 10C12.7761 10 13 10.2239 13 10.5V12C13 13.1046 12.1062 14 11.0012 14H3.99635C2.89019 14 2 13.103 2 12V10.5C2 10.2239 2.22386 10 2.5 10Z'
-        fill={color}
-        fillRule='evenodd'
-        clipRule='evenodd'
-      />
-    </svg>
-  );
-}
-
-export function Window({ size = 14, color = 'currentColor', className = 'inline', ...props }: IconProps) {
-  return (
-    <svg
-      {...props}
-      width={size}
-      height={size}
-      className={className}
-      viewBox='0 0 15 15'
-      xmlns='http://www.w3.org/2000/svg'
-    >
-      <path
-        d='M2.5 1C1.67157 1 1 1.67157 1 2.5V12.5C1 13.3284 1.67157 14 2.5 14H12.5C13.3284 14 14 13.3284 14 12.5V2.5C14 1.67157 13.3284 1 12.5 1H2.5ZM2 2.5C2 2.22386 2.22386 2 2.5 2H12.5C12.7761 2 13 2.22386 13 2.5V4H2V2.5ZM2 5H13V12.5C13 12.7761 12.7761 13 12.5 13H2.5C2.22386 13 2 12.7761 2 12.5V5Z'
-        fill={color}
-        fillRule='evenodd'
-        clipRule='evenodd'
-      />
-    </svg>
-  );
-}
-
-export function Dash({ size = 14, color = 'currentColor', className = 'inline', ...props }: IconProps) {
-  return (
-    <svg
-      {...props}
-      width={size}
-      height={size}
-      className={className}
-      viewBox='0 0 15 15'
-      xmlns='http://www.w3.org/2000/svg'
-    >
-      <path
-        d='M2.25 7.5C2.25 7.08579 2.58579 6.75 3 6.75H12C12.4142 6.75 12.75 7.08579 12.75 7.5C12.75 7.91421 12.4142 8.25 12 8.25H3C2.58579 8.25 2.25 7.91421 2.25 7.5Z'
-        fill={color}
-        fillRule='evenodd'
-        clipRule='evenodd'
-      />
-    </svg>
-  );
-}
