+++ lib/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@moondreamsdev/dreamer-ui",
-  "version": "1.7.9",
+  "version": "1.7.10-test.3",
   "publishConfig": {
     "access": "public"
   },
