+++ lib/src/components/index.ts
@@ -4,6 +4,7 @@ export * from './button';
 export * from './carousel';
 export * from './checkbox';
 export * from './clickable';
+export * from './code-block';
 export * from './input';
 export * from './label';
 export * from './modal';
