+++ lib/src/symbols/Copy.tsx
@@ -1,0 +1,21 @@
+import { SymbolProps } from './props';
+
+export function Copy({ size = 15, color = 'currentColor', className = 'inline', ...props }: SymbolProps) {
+  return (
+    <svg
+      {...props}
+      width={size}
+      height={size}
+      className={className}
+      viewBox='0 0 15 15'
+      xmlns='http://www.w3.org/2000/svg'
+    >
+      <path
+        d='M1 9.50006C1 10.3285 1.67157 11.0001 2.5 11.0001H4L4 10.0001H2.5C2.22386 10.0001 2 9.7762 2 9.50006L2 2.50006C2 2.22392 2.22386 2.00006 2.5 2.00006L9.5 2.00006C9.77614 2.00006 10 2.22392 10 2.50006V4.00002L11 4.00002V2.50006C11 1.67163 10.3284 1.00006 9.5 1.00006H2.5C1.67157 1.00006 1 1.67163 1 2.50006V9.50006ZM5.5 4.00006C4.67157 4.00006 4 4.67163 4 5.50006V12.5001C4 13.3285 4.67157 14.0001 5.5 14.0001H12.5C13.3284 14.0001 14 13.3285 14 12.5001V5.50006C14 4.67163 13.3284 4.00006 12.5 4.00006H5.5ZM5 5.50006C5 5.22392 5.22386 5.00006 5.5 5.00006H12.5C12.7761 5.00006 13 5.22392 13 5.50006V12.5001C13 12.7762 12.7761 13.0001 12.5 13.0001H5.5C5.22386 13.0001 5 12.7762 5 12.5001V5.50006Z'
+        fill={color}
+        fillRule='evenodd'
+        clipRule='evenodd'
+      />
+    </svg>
+  );
+}
