+++ lib/src/symbols/Window.tsx
@@ -1,0 +1,21 @@
+import { SymbolProps } from './props';
+
+export function Window({ size = 15, color = 'currentColor', className = 'inline', ...props }: SymbolProps) {
+  return (
+    <svg
+      {...props}
+      width={size}
+      height={size}
+      className={className}
+      viewBox='0 0 15 15'
+      xmlns='http://www.w3.org/2000/svg'
+    >
+      <path
+        d='M2.5 1C1.67157 1 1 1.67157 1 2.5V12.5C1 13.3284 1.67157 14 2.5 14H12.5C13.3284 14 14 13.3284 14 12.5V2.5C14 1.67157 13.3284 1 12.5 1H2.5ZM2 2.5C2 2.22386 2.22386 2 2.5 2H12.5C12.7761 2 13 2.22386 13 2.5V4H2V2.5ZM2 5H13V12.5C13 12.7761 12.7761 13 12.5 13H2.5C2.22386 13 2 12.7761 2 12.5V5Z'
+        fill={color}
+        fillRule='evenodd'
+        clipRule='evenodd'
+      />
+    </svg>
+  );
+}
