+++ lib/src/utils/index.ts
@@ -1,1 +1,2 @@
+export { downloadFile } from './downloadFile';
 export { join } from './join';
